Add files via upload
In India, farmers have huge farming land and it becomes very difficult for them to reach and track
each corner of such lands. Most of the times there is a possibility of uneven water sprinkles due to some
elevations & slopes. This result in the bad quality crops which further leads to financial losses. In this case the
Smart Irrigation System using IoT for Agriculture 4.0 is very helpful, promising concept which leads to ease of
farming.
The Smart Irrigation System presents a vast opportunity for the automation of the entire irrigation
process. This project involves constructing an IoT-based irrigation system using the ESP8266 NodeMCU Module
and several atmospheric sensors. This system not only has the capability to water the soil automatically based on
its moisture level but also sends the collected data to the Cloud Server, enabling monitoring of the land's
condition. It also includes a water pump that disperses water onto the land, taking into account environmental
factors like moisture, temperature, and humidity.